requirements.txt
streamlit
python-dotenv
requests
pandas
plotly
groq
